Add pixmap_path to the gtkrc. Move install actions from src/ to gui/
The pixmap_path is needed for some engines - they print a warning even though
they eventually find the right pixmaps. This clutters the output and is generally
a poor practice. Plus there's a possibility of incorrect pixmaps being loaded.

The install/uninstall actions related to the GUI are now moved to gui/
directory's Makefile.am - it's cleaner and may marginally reduce install time
when only execuables are being built.

The gui/ install action needs the calfmakerdf program to be present, so there
may be some problems with multithreaded build - I'll check it later.