Separate session manager interface code into separate classes.
All communication with the session manager (lashd) is now done using
a set of interfaces. The code specific for LASH 0.5.4 and LASH 0.6
is now in separate classes. Conditional compilation is used to build
one or the other. There are no more direct references to LASH APIs
from the core JACK host code.