Attempt to fix the Rosegarden crash.
Rosegarden seems to unload the plugins after discovery, which interferes with
the cleanup of local static objects (like one created in ladspa_wrapper::get).
This workaround causes the wrapper object to never be cleaned up, which causes
a memory leak, but fixes the crash. Obviously, if this works, I will have to
convert it into a proper fix using __attribute__((destructor)).