Persist extension 0.2
From the original mail:

Well, it seems clear the subject parameter is a can of worms. I have
removed it. It is still possible to express resources as objects by
using a specific value type for them (e.g. a dictionary). This is more
inline with how implementations would work with such things anyway (and
there are Atom types for this).