Framework: attempt at avoiding input buffers that contain NaN/inf/+192dBFS signals
The first time such a value is detected, a message is printed. Every time
such a value is detected, the buffer is not processed and the output is
forced to all-zeros.

This is to avoid plugins like filter or reverb getting 'stuck' with high
volume signals or +inf/-inf in their state.