Emphasis: add a different version of the FM emphasis
This is based on a high shelf filter fit to a theoretical analog filter
at three points:
- Nyquist frequency
- frequency that corresponds to square root of the gain at Nyquist
- cutoff frequency (corresponding to the original tau)

The advantage of this approach is that it avoids the infinite gain at
Nyquist, which I think may be causing numerical issues with floating
point representation, even if it's followed by another filter that
has a zero at Nyquist.

The main disadvantage is that it's no longer a straightforward digital
version of a single-pole analog filter. The Q calculation for the
shelf filter is based on manual curve-fitting - this should give
good enough results for the typical sample rates, but the cutoff
frequency isn't guaranteed to be spot on. That being said, no digital
filter can have a frequency response that is exactly like the analog
counterpart, for various reasons - aliasing/frequency warping.