Emphasis: fix numerical stability issue in Production mode
In the original implementation, the pole-at-sr/2 highpass filter was followed
by a lowpass filter that removed the infinite gain at Nyquist. Which,
in perfect world would work OK, but fails badly when used with
floating point numbers. The highpass filter keeps accumulating a
DC offset in its coefficients (w1/w2 in Direct II form), which causes
progressive loss of resolution due to limited size of the mantissa
and resulting distortion of the signal. I think the fact that we're
using a Direct II form here makes it even worse.

My proposed solution is to move the lowpass filter *before* the
highpass filter, so that the frequency content near Nyquist is
removed before it can cause a resolution-killing offset in the
highpass that follows it. A cursory check of the filter
state seems to indicate that the problem should be gone or at
least severely diminished.