Limiter: Allow toggling auto-leveling
Sometimes the automatic scaling to 0 dBFS is not desired, for instance
if you want to use the limiter to simply impose a loudness cap on the signal
without changing the overall volume level.

It's currently possible to accomplish this by simply adjusting the
output gain knob to the inverse of the limiter knob, but this can be a
hassle when you want to make quick adjustments.

This patch introduces a new option to disable the auto-leveling
altogether, defaulting to on to avoid changing the current behavior.