Fixup compiler error
Fixup from c41d8c0d033d886df3de8d419edfd1c2a942170f .
That commit tried to pass the static variable `_synths` to a function
(`std::max`), but there is no definition of `_synths`, so it cannot be
passed to a function. However, a simple ternary expression counts as
`constexpr`, not requiring a definition of `_synths`.